file handling improved

+++ safetensors-2-torch/convert_hf_to_torch.py
@@ -38,8 +38,9 @@ def main():
 
     model = torch.load(output_path,weights_only=False)
 
-    with open("model.txt","w") as file:
-        file.write(str(model))
+    if not os.path.isfile("model.txt"):
+        with open("model.txt","w") as file:
+            file.write(str(model))
 
     print("Export Completed!")
